NPC's added. Attempt at animating Dr harris
I added a few NPC's with dialogue boxes that pop u when you get close to them. The dialogue needs to be written up and the box display needs to be adjusted. Hit the space bar to exit conversation. Also added Madam Harris to initial scene and tried to make her the playable character but I could not get the animation and controls to work. Any input is appreciated.