Taco(rabbit) character works as both NPC and inventory item
It works. I edited the script for inventory items to function based on collisions with 2D Box Colliders. The out of the box script used capsule colliders and triggers instead of collisions. This is the only way I could get the NPC controller script to work on the same object as the inventory system. This means that all pre-fabbed inventory items that came with creator kit will not function out of the box any more. Existing pre-fab inventory ites, and any new ones, are pretty easy to make work with the edited script. Add a 2dBoxCollider, leave "is trigger" box UNchecked, and add the inventory item script as a component. For modifying existing pre-fabs remove the capsule collider as well. I have no idea what the long term impact of the script modification might be, but it appears to function just fine right now.