Location adjustments. Modify particle system effect on test tubes
Moved scene tansition "to home" slightly to the right of the main building in covidia. It was kind of in the way of getting to the old woman in the desert. Moved old woman. Adjusted play speed of particle animations on snake oil and disinfectant testubes.  Made magic chicken bigger and expanded magic effect.